Fix presign file upload for user avatar
